requirements.txt
streamlit
google-generativeai
python-docx
openpyxl
pandas
python-dotenv
